% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grassland_05_poly}
\alias{grassland_05_poly}
\title{grassland}
\format{
A data frame with 2749 rows and 3 variables:
\describe{
\item{\code{biotoptype}}{character biotope type}
\item{\code{name_class_eng}}{character biotope text (english)}
\item{\code{geom}}{list geometry column}
}
}
\usage{
grassland_05_poly
}
\description{
sf object containing areas of grassland (biotope type 05)
}
\details{
https://metaver.de/trefferanzeige?docuuid=B57B9F35-AFFF-49F2-BA32-618D1A1CD412
}
\keyword{datasets}
