% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peatlands_04_poly}
\alias{peatlands_04_poly}
\title{swamp & wetland}
\format{
A data frame with 214 rows and 3 variables:
\describe{
\item{\code{biotoptype}}{character biotope type}
\item{\code{name_class_eng}}{character biotope text (english)}
\item{\code{geom}}{list geometry column}
}
}
\usage{
peatlands_04_poly
}
\description{
sf object containing areas of swamps & wetland (biotope type 04)
}
\details{
https://metaver.de/trefferanzeige?docuuid=B57B9F35-AFFF-49F2-BA32-618D1A1CD412
}
\keyword{datasets}
