% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{river_01_poly}
\alias{river_01_poly}
\title{Running water courses polygons}
\format{
A data frame with 161 rows and 3 variables:
\describe{
\item{\code{biotoptype}}{character biotope type}
\item{\code{name_class_eng}}{character biotope text (english)}
\item{\code{geom}}{list geometry column}
}
}
\usage{
river_01_poly
}
\description{
sf object containing running water courses (biotope type 01)
}
\details{
https://metaver.de/trefferanzeige?docuuid=B57B9F35-AFFF-49F2-BA32-618D1A1CD412
}
\keyword{datasets}
