% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_grosskreutz}
\alias{station_grosskreutz}
\title{station_grosskreutz}
\format{
a dataframe with 16 rows and 19 columns
\describe{
\item{\code{name}}{character Station name}
\item{\code{x}}{numeric x coordinate}
\item{\code{y}}{numeric y coordinate}
\item{\code{geometry}}{list geometry column}
}
}
\usage{
station_grosskreutz
}
\description{
sf object with base station location
}
\details{
DETAILS
}
\keyword{datasets}
